Remove the requirement for strict status checks
This check requires that a branch is up to date
wrt. the default branch before a PR can be merged.
This doesn't apply for, e.g. the cloud-platform-
environments repository, or any other repo. where
multiple, non-overlapping changes are likely to be
made.

Arguably, this shouldn't be one of our standard
requirements for our repositories, so this change
removes it altogether.